pyproject.toml
[project]
name = "anifeatures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.9.0",
    "imagehash>=4.3.2",
    "open-clip-torch>=3.3.0",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pillow-avif-plugin>=1.5.5",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
]
